+++ CHANGELOG.md
@@ -4,11 +4,17 @@
 ### Added
 - File structure and package.json
 
-## [1.0.1] - 2020.20.13.
+### Removed
+- All unnecessary meta files
+
+## [1.0.1] - 2020.10.13.
 ### Added
 - IEntity 
 - Command for the Command Pattern
 - CommandProcessor that handles all Commands
 
-### Removed
-- All unnecessary meta files
+
+## [1.0.2] - 2020.10.19.
+### Added
+- Singleton Pattern
+- Generic Singleton that derives from MonoBehaviour both persistent and non